Move getThemeFilterStringFromTerm
